move password to ui package & ui fixes
